+++ gamemodes/nzombies/entities/entities/nz_zombie_walker.lua
@@ -384,11 +384,11 @@ function ENT:TriggerBarricadeJump()
 		local seqtbl = self.ActStages[self:GetActStage()] and self[self.ActStages[self:GetActStage()].barricadejumps] or self.JumpSequences
 		local seq = seqtbl[math.random(#seqtbl)]
 		local id, dur = self:LookupSequence(seq.seq)
-		self:SetSolidMask(MASK_NPCSOLID_BRUSHONLY)
-		self:SetCollisionGroup(COLLISION_GROUP_DEBRIS_TRIGGER)
+		self:SetSolidMask(MASK_SOLID_BRUSHONLY)
+		--self:SetCollisionGroup(COLLISION_GROUP_DEBRIS_TRIGGER)
 		self.loco:SetAcceleration( 5000 )
 		self.loco:SetDesiredSpeed(seq.speed)
-		self:SetVelocity(self:GetForward()*seq.speed)
+		self:SetVelocity(self:GetForward() * seq.speed)
 		self:SetSequence(id)
 		self:SetCycle(0)
 		self:SetPlaybackRate(1)
